+++ back/src/KI/CoreBundle/Resources/config/entities.yml
@@ -25,6 +25,7 @@ parameters:
             - Event
             - Exercice
             - Newsitem
+            - Message
         user:
             - Club
             - Experience
