+++ back/src/KI/PublicationBundle/Entity/NewsitemFile.php
@@ -13,7 +13,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
  * @ORM\HasLifecycleCallbacks
  * @JMS\ExclusionPolicy("all")
  */
-class PostFile
+class NewsitemFile
 {
     /**
      * @ORM\Id
@@ -49,11 +49,11 @@ class PostFile
     protected $file;
 
     /**
-     * @var \KI\PublicationBundle\Entity\Post
-     * @ORM\ManyToOne(targetEntity="KI\PublicationBundle\Entity\Post", inversedBy="files")
-     * @ORM\JoinColumn(name="post_id", referencedColumnName="id")
+     * @var \KI\PublicationBundle\Entity\Newsitem
+     * @ORM\ManyToOne(targetEntity="KI\PublicationBundle\Entity\Newsitem", inversedBy="files")
+     * @ORM\JoinColumn(name="newsitem_id", referencedColumnName="id")
      **/
-    private $post;
+    private $newsitem;
 
     /**
      * @return string
@@ -68,7 +68,7 @@ class PostFile
      */
     protected function getUploadCategory()
     {
-        return 'posts';
+        return 'newsitems';
     }
 
 
@@ -91,7 +91,7 @@ class PostFile
      */
     public function url()
     {
-        return 'uploads/'.$this->getUploadCategory().'/'.$this->post->getId()."_".$this->getName();
+        return 'uploads/'.$this->getUploadCategory().'/'.$this->newsitem->getId()."_".$this->getName();
     }
 
     /**
@@ -108,7 +108,7 @@ class PostFile
      * Set ext
      *
      * @param string $ext
-     * @return PostFile
+     * @return NewsitemFile
      */
     public function setExt($ext)
     {
@@ -131,7 +131,7 @@ class PostFile
      * Set name
      *
      * @param string $name
-     * @return PostFile
+     * @return NewsitemFile
      */
     public function setName($name)
     {
@@ -154,7 +154,7 @@ class PostFile
      * Set size
      *
      * @param integer $size
-     * @return PostFile
+     * @return NewsitemFile
      */
     public function setSize($size)
     {
@@ -191,7 +191,7 @@ class PostFile
      * @ORM\PostPersist()
      */
     public function moveFile(){
-        $this->file->move($this->getUploadDir(), $this->post->getId()."_".$this->getName());
+        $this->file->move($this->getUploadDir(), $this->newsitem->getId()."_".$this->getName());
     }
 
     /**
@@ -208,16 +208,16 @@ class PostFile
     /**
      * @return mixed
      */
-    public function getPost()
+    public function getNewsitem()
     {
-        return $this->post;
+        return $this->newsitem;
     }
 
     /**
-     * @param mixed $post
+     * @param mixed $newsitem
      */
-    public function setPost($post)
+    public function setNewsitem($newsitem)
     {
-        $this->post = $post;
+        $this->newsitem = $newsitem;
     }
 }
