+++ back/src/KI/PublicationBundle/Form/MessageType.php
@@ -6,27 +6,20 @@ use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class NewsitemType extends AbstractType
+class MessageType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('name')
             ->add('text')
-            ->add('authorClub', 'club_selector')
-            ->add('uploadedFiles', 'file', array(
-                    'multiple' => true,
-                    'data_class' => null,
-                    'required' => false,
-                )
-            )
-        ;
+            ->add('image', 'image_selector')
+          ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'data_class' => 'KI\PublicationBundle\Entity\Newsitem',
+            'data_class' => 'KI\PublicationBundle\Entity\Message',
             'csrf_protection' => false,
         ));
     }
